refactor : redesign common package
