+++ src/main/java/coz/weavon/common/io/ResponseType.java
@@ -1,4 +1,4 @@
-package coz.weavon.core.shared.presentation.model.response;
+package coz.weavon.common.io;
 
 public enum ResponseType {
     SUCCESS,
