+++ src/main/java/coz/weavon/common/validator/ValidRange.java
@@ -1,6 +1,5 @@
-package coz.weavon.validation.model;
+package coz.weavon.common.validator;
 
-import coz.weavon.validation.validator.RangeValidator;
 import jakarta.validation.Constraint;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
