+++ src/main/java/coz/weavon/core/auth/domain/service/OAuthUserProvider.java
@@ -1,7 +1,7 @@
 package coz.weavon.core.auth.domain.service;
 
+import coz.weavon.common.exception.BusinessException;
 import coz.weavon.core.auth.domain.model.OAuthUser;
-import coz.weavon.exception.model.BusinessException;
 import java.util.Map;
 import lombok.experimental.UtilityClass;
