+++ src/main/java/coz/weavon/core/project/application/model/command/ProjectSearchCommand.java
@@ -1,7 +1,7 @@
 package coz.weavon.core.project.application.model.command;
 
+import coz.weavon.common.shared.Command;
 import coz.weavon.core.project.domain.model.ProjectMemberRole;
-import coz.weavon.core.shared.application.model.command.RestCommand;
 import java.util.Collections;
 import java.util.List;
 import lombok.Builder;
@@ -10,7 +10,7 @@ import org.springframework.util.CollectionUtils;
 
 @Getter
 @Builder
-public class ProjectSearchCommand extends RestCommand {
+public class ProjectSearchCommand extends Command {
 
     private List<Long> projectIds;
