+++ src/main/java/coz/weavon/core/user/application/model/command/UserSearchCommand.java
@@ -1,13 +1,13 @@
 package coz.weavon.core.user.application.model.command;
 
-import coz.weavon.core.shared.application.model.command.RestCommand;
+import coz.weavon.common.shared.Command;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class UserSearchCommand extends RestCommand {
+public class UserSearchCommand extends Command {
 
     private List<Long> userIds;
