+++ src/main/java/coz/weavon/core/user/application/service/UserRestService.java
@@ -1,12 +1,12 @@
 package coz.weavon.core.user.application.service;
 
+import coz.weavon.common.exception.BusinessException;
 import coz.weavon.core.user.application.model.command.UserOperateCommand;
 import coz.weavon.core.user.application.model.command.UserSearchCommand;
 import coz.weavon.core.user.application.model.result.UserOperateResult;
 import coz.weavon.core.user.application.repository.UserRepository;
 import coz.weavon.core.user.domain.model.User;
 import coz.weavon.core.user.domain.model.Users;
-import coz.weavon.exception.model.BusinessException;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
