+++ src/main/java/coz/weavon/security/filter/UsernameAuthenticationFilter.java
@@ -1,10 +1,10 @@
 package coz.weavon.security.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import coz.weavon.common.io.ErrorResponse;
+import coz.weavon.common.io.RestResponse;
 import coz.weavon.core.auth.domain.model.AuthToken;
 import coz.weavon.core.auth.domain.model.AuthUser;
-import coz.weavon.core.shared.presentation.model.response.ErrorResponse;
-import coz.weavon.core.shared.presentation.model.response.RestResponse;
 import coz.weavon.helper.MessageTranslator;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.Cookie;
